It will be useful for us to define a default difficulty as a value: default : Difficulty default =     Any
Also, we will need a list of all of the difficulties, and the problem with union types is that you cannot get all of the values. So I need to define a list that will contain all of these, so I can iterate over them and, for example, display them on the screen, so I'm gonna go ahead and put all of them into a list here:
list =
    [ Any
    , Easy
    , Medium
    , Hard
    ]
... and now I have a list of type Difficulty:
list : List Difficulty
list =
    [ Any
    , Easy
    , Medium
    , Hard
    ]

... and now I can use it but it's not super helpful in our case. What would be better is have a string representation of all of these difficulties. So what I'm gonna do is, I'm gonna have a tuple, which will have a string associated with each of the difficulties:
list =
    [ ("Any", Any)
    , ("Easy", Easy)
    , ("Medium", Medium)
    , ("Hard", Hard)
    ]
... so I'm displaying an "Any" string for Any Difficulty, and "Easy" string for Easy Difficulty, etc.
... This will throw an arror, because we still need to adjust the type definition on our list:
`The definition of `list` does not match its type annotation`. So in the next commit we'll change the type signature to fix this error.