Let's go ahead and define a new view for our question that we can use in our application. Let's get rid of all of that. I'll use let ... in statement to calculate the list of all the available answers so that users will be able to choose from those: view : Question -> String view { question, correct, incorrect } =     let         answers =             correct :: incorrect     in         div [] [] ...and we'll put the correct list on top of incorrect list: `correct :: incorrect`
Now there is a problem, cuz the first answer will always be the correct one.
So let's go ahead and sort that list, so that we have a consistent output of questions and answers.

Now we can take these answers, go into the descendants, and let's map over those answers, and just create a text node out of them.

I forgot to import div from Html package, in Question.elm. And also text, and the Html type, because now this view function outputs Html. And the return value of this application will be a list, so:
div []
    ( answers
        |> List.map text
    )
... we need to wrap the above into parens () to tell our compiler that this is a list, so we're passing (...) as a second argument, cuz that will get mapped into a List.